Various games created using Python 3.
Guessing game - In this game, the program chooses a random number and the  player has to guess the number. As the player gets closer to predicting the number, the program outputs that its getting "warmer" and as the player gets father away from the chosen number, the program outputs that it's getting "colder". finally if the prediction is correct, it outputs so.

Black Jack - Get as close to 21 as you can without going over! Dealer hits until she reaches 17. Aces count as 1 or 11.

Card War - I'm using Python OOP to simulate a simplified version of the game war. Two players will each start off with half the deck, then they each remove a card, compare which card has the highest value, and the player with the higher card wins both cards.